Rest of entities

+++ LMSAdmin/src/main/java/com/lms/LMSAdmin/controller/AuthorController.java
@@ -1,10 +1,10 @@
 package com.lms.LMSAdmin.controller;
 
-import java.sql.SQLException;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -15,54 +15,50 @@ import com.lms.LMSAdmin.pojo.Author;
 import com.lms.LMSAdmin.service.AuthorService;
 
 @RestController
+@RequestMapping("/lmsAdmin/menu/authMenu")
 public class AuthorController {
 	
 	@Autowired
 	AuthorService authorService;
 	
-	//test
-	@RequestMapping("/lmsAdmin/author/{authorId}")
-	public Author getAuthorById(@PathVariable int authorId) throws SQLException {
-		return authorService.getAuthorById(authorId);
-	}
-	
-	/*@RequestMapping("/lmsAdmin/menu/1/authMenu/{authOpt}")
-	public void authorMenu(@PathVariable int authOpt) throws SQLException {
-		switch(authOpt) {
-			case 1:
-				insertAuthor();
- 			case 4: 
-				getAllAuthors();
-				break;
-			default:
-				
-				break;
-		}
-	}*/
-	
 	//Create a record
-	@RequestMapping(value = "/lmsAdmin/menu/2/authMenu/1/authName/{authorName}", method = {RequestMethod.POST, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.CREATED, reason = "Author record created.")
-	public void insertAuthor(@PathVariable String authorName) {
+	@RequestMapping(value = "/create/authName/{authorName}", method = {RequestMethod.POST, RequestMethod.GET})
+	public ResponseEntity<String> insertAuthor(@PathVariable String authorName) {
+		
 		authorService.insertAuthor(authorName);
+		return new ResponseEntity<String>("Author record created.", HttpStatus.CREATED);
 	}
 	
 	//Update a record
-	@RequestMapping(value = "/lmsAdmin/menu/2/authMenu/2/authId/{authorId}/authName/{authorName}", method = {RequestMethod.PUT, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.OK, reason = "Author record updated.")
-	public void updateAuthor(@PathVariable("authorId") int authorId, @PathVariable("authorName") String authorName) {
-		authorService.updateAuthor(authorId, authorName);
+	@RequestMapping(value = "/update/authId/{authorId}/authName/{authorName}", method = {RequestMethod.PUT, RequestMethod.GET})
+	public ResponseEntity<String> updateAuthor(@PathVariable("authorId") int authorId, @PathVariable("authorName") String authorName) {
+		
+		boolean checkId = authorService.ifExists(authorId);
+		
+		if(checkId == true) {
+			authorService.updateAuthor(authorId, authorName);
+			return new ResponseEntity<String>("Author record updated.", HttpStatus.ACCEPTED);
+		}else {
+			return new ResponseEntity<String>("Invalid ID.", HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	//Delete a record
-	@RequestMapping(value = "/lmsAdmin/menu/2/authMenu/3/authId/{authorId}", method = {RequestMethod.DELETE, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.NO_CONTENT, reason = "Author record deleted.")
-	public void deleteAuthor(@PathVariable int authorId) {
-		authorService.deleteAuthor(authorId);
+	@RequestMapping(value = "/delete/authId/{authorId}", method = {RequestMethod.DELETE, RequestMethod.GET})
+	public ResponseEntity<String> deleteAuthor(@PathVariable int authorId) {
+		
+		boolean checkId = authorService.ifExists(authorId);
+		
+		if(checkId == true) {
+			authorService.deleteAuthor(authorId);
+			return new ResponseEntity<String>("Author record deleted.", HttpStatus.NO_CONTENT);
+		}else {
+			return new ResponseEntity<String>("Invalid ID.", HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	//View all records
-	@RequestMapping(value = "/lmsAdmin/menu/2/authMenu/4", method = RequestMethod.GET)
+	@RequestMapping(value = "/view", method = RequestMethod.GET)
 	@ResponseStatus(code = HttpStatus.OK)
 	public List<Author> getAllAuthors() {
 		return authorService.getAllAuthors();
